refactor and code cleaning
