+++ .gitignore
@@ -262,3 +262,5 @@ dmypy.json
 
 # Cython debug symbols
 cython_debug/
+
+.ruff_cache/
