+++ app.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from src.content import (
+from src.config.content import (
     HERO_TEXT,
     ABOUT_TEXT,
     CITATION_LABEL,
@@ -10,13 +10,13 @@ from src.content import (
     SUPPORT_TEXT,
 )
 
-from src.expert import expert_mode
-from src.calculator import calculator_mode
-from src.token_estimator import token_estimator
+from src.ui.expert import expert_mode
+from src.ui.calculator import calculator_mode
+from src.ui.token_estimator import token_estimator
 
 st.set_page_config(layout="wide", page_title="EcoLogits Calculator", page_icon="🧮")
 
-with open("src/style.css") as css:
+with open("src/ui/style.css") as css:
     st.markdown(f"<style>{css.read()}</style>", unsafe_allow_html=True)
 
 st.html(HERO_TEXT)
