+++ pyproject.toml
@@ -8,6 +8,7 @@ dependencies = [
     "ecologits>=0.9.0,<0.10.0",
     "pint>=0.24.4",
     "plotly>=6.2.0",
+    "ruff>=0.14.7",
     "streamlit>=1.47.1",
     "tiktoken>=0.9.0",
 ]
